Add ScheduledItem.objects.available() method and new ScheduledItemQuerySet
This method grabs the current list of available schedule objects and returns
the QuerySet.  You can filter it further by adding additional filter() calls
before evaluating it.

The ScheduledItemQuerySet allows you to reach through the ScheduledItem and
return the objects that relate to it by adding .return_related() to the
QuerySet method calls.